--- curl.py
@@ -1,25 +1,0 @@
-import json
-import os
-import time
-
-for i in range(1):
-
-    # 定義要發送的資料
-    data = {'id': '6703545e53666b0012e8bc33', 'formId': 'Wo4pX5iW', 'formName': 'Counterspell Taiwan 加入我們 (系統測試)', 'fields': [{'title': ['全台首次！（測試版）', ['br'], '青少年為青少年舉辦的黑客松，等你來加入！💡'], 'description': [['div', ['我們是誰？&nbsp;']], ['div', [['br']]], ['div', ['我們是一群來自台灣的高中生，正在籌備全台灣第一場由青少年舉辦、專為青少年設計的黑客松活動 —— Counterspell Taiwan！🎉 這場活動充滿挑戰和創意，不僅限於技術領域，無論你對程式開發、活動策劃、接洽、攝影、美術設計感興趣，我們都非常期待你的加入！這是屬於我們高中生的舞台，一個我們可以共同協作、突破創新的絕佳機會。', ['br']]], ['div', [['div', [['br']]], ['div', [['b', ['我們正在做什麼？&nbsp;']]]], ['div', [['br']]], ['div', ['我們，來自台灣的高中自學生，正在與全球知名技術社群 Hack Club 共同籌辦一場30小時的黑客松！🎉Hack Club 是一個全球高中生組成的開源社群，旨在透過程式開發和項目合作，讓年輕人學習技術、實踐創意。無論你是技術新手還是資深開發者，Hack Club 都是一個讓你自由創作、學習成長的平台！想了解更多？請點擊這裡了解&nbsp;', ['a', ['Hack Club'], {'href': 'https://hackclub.com/'}], '。&nbsp;']], ['div', [['br']]], ['div', ['這場 30 小時的黑客松將於2024年11月30日至12月01日在台北市舉行，參加者將在這段時間內組隊合作，從構思到開發，完成一個有趣且實用的項目。我們致力於創造一個充滿創意、挑戰和樂趣的環境，讓每位參加者都能享受創作的過程，並展示出讓人驚艷的成果！', ['br']]], ['div', [['br']]], ['div', [['b', ['我們需要你！&nbsp;']]]], ['div', [['br']]], ['div', ['我們的黑客松需要來自各領域的合作夥伴！不論你是否懂程式設計，還是對活動策劃、接洽、視覺設計、攝影等領域感興趣，我們都希望你能加入我們的團隊，貢獻你的才華和創意！即使你沒有技術背景，只要你對這場活動充滿熱情和好奇，我們也非常期待你的加入！', ['br']]], ['div', [['br']]], ['div', [['b', ['有其他想法或建議？加入我們！&nbsp;']]]], ['div', [['br']]], ['div', ['這場黑客松不僅是我們的，也是你的！如果你對活動有任何其他建議或想法，也歡迎你加入我們，讓這次活動變得更加精彩！']]]]], 'kind': 'welcome', 'validations': None, 'properties': {'buttonText': '開始填寫!'}, 'id': 'KyS99WqdzDdu', 'layout': None}, {'title': ['關於 Counterspell\n\n'], 'description': [['div', ['Counterspell 是什麼？', ['br']]], ['div', [['br']]], 'Counterspell 是由 Hack Club 所發起的全球最大規模的同步性活動之一。這是一場 30 小時的遊戲創作馬拉松，專為高中生設計，旨在讓程式開發者、藝術家和音樂家齊聚一堂，共同合作完成遊戲創作。這次活動將於 11 月底在臺北市舉行，並且與全球 200 多個地點同步進行（原定於 11/23、11/24 舉辦，但逢段考前週末，臺灣活動將延至 11/30、12/1 舉辦）&nbsp;', ['div', [['br']]], ['div', ['遊戲創作馬拉松（Game Jam）是黑客松（Hackathon）的一種形式，參與者可以透過這樣的活動提升他們的創意和技能，並與來自不同領域的夥伴合作。這是一個友好的環境，特別適合初學者，他們可以在這裡學習、探索遊戲開發的過程，並激發自己的創新潛力。&nbsp;&nbsp;', ['br']]], ['div', [['br']]], ['div', ['本次活動由 Hack Club 協辦，他們將提供全方位的支援、資源以及全球性的參與體驗。所有參與者都將獲得一件全球同步的 Counterspell T-shirt，同時，工作人員還有機會獲得專屬的工作人員 T-shirt。Hack Club 是一個美國 501(c)(3) 非營利組織，致力於透過全球範圍內的程式開發社群和活動，幫助超過 30,000 名青少年進入資訊領域。他們曾舉辦並支持無數的黑客松活動，其中包括在', ['a', ['穿越加拿大的火車上'], {'href': 'https://www.youtube.com/watch?v=hiG3fYq3xUU'}], '舉辦的黑客松，以及在', ['a', ['荒野中的黑客松'], {'href': 'https://www.youtube.com/watch?v=ufMUJ9D1fi8'}], '活動。這些活動不僅有趣，更激發了全球青少年的創作熱情，真的很酷，對吧？&nbsp;&nbsp;', ['br']]], ['div', [['br']]], ['div', ['我們的願景！']], ['div', [['br']]], ['div', ['Counterspell 將是台灣有史以來第一場由青少年為青少年舉辦的黑客松活動。然而，在其他先進國家（如美國、歐洲），這樣的活動早已行之有年。我們希望透過這次活動，讓台灣社會看見並了解青少年有能力主導大型創意項目，並引領這樣的風氣在台灣持續發展。因此，我們成立了一個組織——HackIt，我們將引領台灣青少年的創新與科技教育。']], ['div', [['br']]], ['div', ['同時，我們還計劃在明年寒假舉辦一場 GenAI 黑客松 活動！如果您對此感興趣，請加入我們！協助我們的願景成真，讓更多人看到青少年的無限可能。']], ['div', [['br']]], ['div', ['加入我們的理由！', ['br']]], ['div', [['br']]], ['div', ['我們的團隊非常強大，集合了各方的專業人才，以確保活動的萬無一失。請想想，一生中有多少次機會是有人主動提供資源，邀請您一起創造一個令人難忘的活動？這是一次難得的機會，我們誠摯地邀請您參與，與我們一起推動台灣的創新未來。']], ['div', [['br']]], ['div', ['您的加入將是實現這個願景的重要力量。我們非常感謝您的任何形式的支持，並期待與您攜手，為這次盛會增添光彩。', ['br']]]], 'kind': 'statement', 'validations': None, 'properties': None, 'id': '6DSIRzRRQhh0', 'layout': None}, {'title': ['你的名字?'], 'description': [], 'kind': 'short_text', 'validations': {'required': True}, 'properties': None, 'id': 'S5jaJPUkBIkt', 'layout': {'mediaType': 'image', 'mediaUrl': 'http://form.xiaobo.tw/static/upload/CmF0gwqTPTmf-DALL%C3%82%C2%B7E%202024-10-03%2014.00.53%20-%20A%20high-end%2C%20clean%20webpage%20design%20with%20a%20floating%20soap%20bubble%20fully%20embodying%20the%20sweetness%20of%20water.%20The%20background%20features%20soft%20ripples%20and%20fluid%2C%20l.png', 'brightness': 0, 'align': 'split_right'}}, {'title': ['電子郵件~'], 'description': [], 'kind': 'email', 'validations': {'required': True}, 'properties': None, 'id': 'ApY1CyoJF6Rv', 'layout': {'mediaType': 'image', 'mediaUrl': 'http://form.xiaobo.tw/static/upload/YLqY2i7FO4Jl-DALL%C3%82%C2%B7E%202024-10-03%2009.45.31%20-%20A%20clean%2C%20minimalistic%20illustration%20featuring%20a%20single%20soap%20bubble%20with%20iridescent%20reflections%2C%20inside%20which%20is%20a%20white%20and%20yellow%20email%20envelope%20symbo.png', 'brightness': 0, 'align': 'split_right'}}, {'title': ['電話號碼~~'], 'description': [], 'kind': 'phone_number', 'validations': {'required': True}, 'properties': {'defaultCountryCode': 'TW'}, 'id': '52lBVg2N0KoU', 'layout': {'mediaType': 'image', 'mediaUrl': 'http://form.xiaobo.tw/static/upload/Y1OnwAEm8QBD-DALL%C3%82%C2%B7E%202024-10-03%2014.02.22%20-%20A%20high-end%2C%20clean%20webpage%20design%20featuring%20a%20large%2C%20smooth%2C%20and%20transparent%20soap%20bubble%20floating%20over%20a%20vast%2C%20serene%20water%20surface.%20Inside%20the%20soap%20bu.png', 'brightness': 0, 'align': 'split_right'}}, {'title': ['高中階段~~'], 'description': [], 'kind': 'multiple_choice', 'validations': {'required': True}, 'properties': {'choices': [{'id': 'FfKX9DzFyPO5', 'label': '高一'}, {'id': 'XLyjhSereJYK', 'label': '高二'}, {'id': 'N3nMDkxAPFVa', 'label': '高三'}, {'id': 'Dm3tO8L5RItg', 'label': '其他'}]}, 'id': 'P7YquMjBAMb6', 'layout': {'mediaType': 'image', 'mediaUrl': 'http://form.xiaobo.tw/static/upload/yVRpClI4PSBD-DALL%C3%82%C2%B7E%202024-10-03%2013.42.57%20-%20A%20minimalist%20design%20featuring%20a%20question%20mark%20shaped%20like%20a%20soap%20bubble%2C%20with%20a%20transparent%2C%20iridescent%20look.%20The%20question%20mark%20floats%20above%20a%20large%2C%20.png', 'brightness': 0, 'align': 'split_right'}}, {'title': ['你住在哪~~~'], 'description': ['請填寫縣市(如:&nbsp; 臺中市)'], 'kind': 'short_text', 'validations': {'required': True}, 'properties': None, 'id': '1yLzipGL8CXc', 'layout': {'mediaType': 'image', 'mediaUrl': 'http://form.xiaobo.tw/static/upload/4TOL0lqpOFA6-DALL%C3%82%C2%B7E%202024-10-03%2013.53.27%20-%20A%20high-end%2C%20clean%20webpage%20design%20featuring%20a%20floating%20soap%20bubble%20with%20an%20abstract%2C%20minimalist%20globe%20inside.%20The%20globe%20is%20represented%20by%20smooth%2C%20curve.png', 'brightness': 0, 'align': 'split_right'}}, {'title': ['來找找適合你的領域~'], 'description': ['無論你擅長什麼領域，我們都有適合你的角色！快看看你可以幫助我們的地方：&nbsp;', ['div', [['br']]], ['div', ['公關組：']], ['div', ['負責與企業、政府、組織和媒體接洽，處理場地租用、舞台設備、文件申請，並進行媒體宣傳等。&nbsp;']], ['div', [['br']]], ['div', ['活動企劃組：']], ['div', ['負責活動流程設計、講座安排、競賽規則與評審安排，並協助參賽者的食宿及餐飲服務。&nbsp;']], ['div', [['br']]], ['div', ['美術組：']], ['div', ['負責設計黑客松的主視覺、海報及社群媒體貼文等。&nbsp;']], ['div', [['br']]], ['div', ['資訊組：']], ['div', ['負責網站、系統開發及相關技術支援，如RDIF投票系統設計等。&nbsp;']], ['div', [['br']]], ['div', ['影音組：']], ['div', ['負責活動紀錄（攝影）、影片剪輯、動畫製作與音樂創作。&nbsp;']], ['div', [['br']]], ['div', ['行政組：（暫不招收）']], ['div', ['處理報名資料、人事管理及相關行政事務。&nbsp;']], ['div', [['br']]], ['div', ['財務組：（暫不招收）']], ['div', ['負責活動的財務管理及報銷事務（不包含募資）。']]], 'kind': 'multiple_choice', 'validations': {'required': True}, 'properties': {'allowMultiple': True, 'choices': [{'id': 'HDDzjo46UCML', 'label': '公關組'}, {'id': 'BryEnikchCiP', 'label': '活動企劃組'}, {'id': 'QTBS6ThhjimL', 'label': '美術組'}, {'id': 'CL3C1UBMmCpy', 'label': '資訊組'}, {'id': 'LMAYGT8Xo4zc', 'label': '影音組'}]}, 'id': 'umigMKtao9y7', 'layout': {'mediaType': 'image', 'mediaUrl': 'http://form.xiaobo.tw/static/upload/ubwJE6GXlenb-DALL%C3%82%C2%B7E%202024-10-03%2014.22.07%20-%20A%20creative%2C%20high-end%20webpage%20design%20featuring%20a%20large%2C%20smooth%2C%20and%20transparent%20soap%20bubble%20floating%20over%20a%20serene%20water%20surface.%20Inside%20this%20large%20bub.png', 'brightness': 0, 'align': 'split_right'}}, {'title': ['讓我們認識一下，好嗎？ 😊'], 'description': ['我們非常期待能與你合作！在進一步了解你的過程中，我們準備了四個問題，讓我們更清楚你是誰、你的興趣和你希望在這次黑客松中發揮的角色。這不僅幫助我們找到最適合你的位置，也讓我們能一起把這場活動變得更精彩！', ['br']], 'kind': 'statement', 'validations': {'required': True}, 'properties': {'buttonText': '開始吧'}, 'id': 'UQfnQOcDcqgB', 'layout': None}, {'title': ['排一排，告訴我們你的優先選擇吧！✨'], 'description': ['(如果你在前面只選擇了一個組別，可以放心的跳過這題!)', ['div', [['div', ['如果你在前面選擇了多個組別，我們想知道你最想參與的是哪一個，請幫我們排一下順序。這樣，我們可以更好地了解你的偏好，並確保你能加入最喜歡的組別。', ['br']]], ['div', [['br']]], ['div', ['如果你忘了剛剛填了什麼組別，這或許可以幫助你回憶一下:', ['br'], '- 公關組&nbsp;']], ['div', ['- 活動企劃組&nbsp;']], ['div', ['- 美術組&nbsp;']], ['div', ['- 資訊組&nbsp;']], ['div', ['- 影音組&nbsp;']], ['div', ['- 行政組&nbsp;']], ['div', ['- 財務組', ['br']]], ['div', ['(如果還是沒記起來，沒關係，可以回到前面看看你的答案~)']], ['div', [['br']]], ['div', ['來吧！用下面的格式列出你的順位：']], ['div', ['(如果前面選了兩個，那這邊就寫第一和第二順位，依此類推😊)&nbsp;']], ['div', ['第一順位：&nbsp;']], ['div', ['第二順位：&nbsp;']]]]], 'kind': 'long_text', 'validations': {'required': False}, 'properties': None, 'id': '3DqecY2ogvR5', 'layout': {'mediaType': 'image', 'mediaUrl': 'http://form.xiaobo.tw/static/upload/8f4MsUyECLkf-DALL%C3%82%C2%B7E%202024-10-03%2014.32.12%20-%20A%20clean%2C%20high-end%20webpage%20design%20featuring%20a%20large%2C%20transparent%20soap%20bubble%20floating%20over%20a%20serene%20water%20surface.%20Inside%20the%20bubble%2C%20there%20is%20only%20a%20s.png', 'brightness': 0, 'align': 'split_right'}}, {'title': ['為什麼選擇這些組別？🎯'], 'description': ['我們超好奇你對這些組別的想法！分享一下吧——什麼讓你對這些組別感興趣？是什麼地方吸引了你呢？你希望在這裡學到什麼新技能，或是實現哪些目標？無論是個人挑戰還是想學習新事物，我們都很期待聽到你的故事！', ['div', [['br']]], ['div', ['這是讓我們更了解你的好機會，來說說你的故事吧！🔥&nbsp;&nbsp;', ['br']]]], 'kind': 'long_text', 'validations': {'required': True}, 'properties': None, 'id': 'fV33uN18Aq5b', 'layout': {'mediaType': 'image', 'mediaUrl': 'http://form.xiaobo.tw/static/upload/DzvmMXJWaF0J-DALL%C3%82%C2%B7E%202024-10-03%2014.32.12%20-%20A%20clean%2C%20high-end%20webpage%20design%20featuring%20a%20large%2C%20transparent%20soap%20bubble%20floating%20over%20a%20serene%20water%20surface.%20Inside%20the%20bubble%2C%20there%20is%20only%20a%20s.png', 'brightness': 0, 'align': 'split_right'}}, {'title': ['有什麼相關經驗或技能嗎？💡'], 'description': ['我們想了解你過去的經驗！如果你曾經參與過相關活動、學過一些技術，或者有任何可以運用在這次黑客松中的技能，都歡迎和我們分享。這些經歷不僅能幫助我們更好地了解你，也能讓你在這次活動中發揮更大的作用！', ['div', [['br']]], ['div', ['不管經驗是多是少，任何分享都對我們非常有幫助，我們期待了解更多關於你的亮點！🌟', ['br']]]], 'kind': 'long_text', 'validations': {'required': True}, 'properties': None, 'id': 'bdXbCUl1iiqq', 'layout': {'mediaType': 'image', 'mediaUrl': 'http://form.xiaobo.tw/static/upload/UDphVwdOqgjz-DALL%C3%82%C2%B7E%202024-10-03%2014.32.12%20-%20A%20clean%2C%20high-end%20webpage%20design%20featuring%20a%20large%2C%20transparent%20soap%20bubble%20floating%20over%20a%20serene%20water%20surface.%20Inside%20the%20bubble%2C%20there%20is%20only%20a%20s.png', 'brightness': 0, 'align': 'split_right'}}, {'title': ['是否同意我們蒐集並使用您的資料？'], 'description': ['感謝您的關注與報名！', ['div', [['br'], ['div', ['由於申請人數眾多，我們特別開發了一套高效的自動化系統來簡化流程。']], ['div', ['您的資料將依照最高安全標準加密保存於我們的資料庫中（相關程式碼已在我們的 HackIt GitHub 組織中開源），並僅限於 HackIt 團隊內部在相關行政用途上使用。\n\n']], ['div', [['br']]], ['div', ['請放心，我們絕不會與任何第三方共享或洩漏您的資料。&nbsp;', ['br']]], ['div', ['若您同意，請在下方簽名。我們將盡速與您聯繫！', ['br']]]]]], 'kind': 'signature', 'validations': {'required': True}, 'properties': {}, 'id': 'Gkg271OtI9Bw', 'layout': None}, {'title': ['非常感謝你填寫這份表單！🙏'], 'description': ['我們很期待和你一起參加這場充滿創意和挑戰的黑客松！如果有任何問題，或需要進一步聯繫，我們會透過 Email 與你聯絡，記得保持信箱暢通哦📬。', ['div', [['br']]], ['div', ['再次感謝你對我們活動的支持，我們非常期待與你見面，共同創造精彩時刻！🎉']], ['div', [['br']]], ['div', ['聯絡我們：&nbsp;']], ['div', ['📸 IG: ', ['a', ['@hackit.tw'], {'href': 'https://www.instagram.com/hackit.tw/'}]]], ['div', ['✉️ EMAIL: counterspell@hackit.tw', ['br']]]], 'kind': 'thank_you', 'validations': None, 'properties': {'redirectOnCompletion': False}, 'id': 'RzBo7OCuX7nc', 'layout': None}], 'answers': [{'id': 'S5jaJPUkBIkt', 'title': '你的名字?', 'kind': 'short_text', 'value': '洪承佑'}, {'id': 'ApY1CyoJF6Rv', 'title': '電子郵件~', 'kind': 'email', 'value': 'xiao.bo.nor@gmail.com'}, {'id': '52lBVg2N0KoU', 'title': '電話號碼~~', 'kind': 'phone_number', 'properties': {'defaultCountryCode': 'TW'}, 'value': '+886937008951'}, {'id': 'P7YquMjBAMb6', 'title': '高中階段~~', 'kind': 'multiple_choice', 'properties': {'choices': [{'id': 'FfKX9DzFyPO5', 'label': '高一'}, {'id': 'XLyjhSereJYK', 'label': '高二'}, {'id': 'N3nMDkxAPFVa', 'label': '高三'}, {'id': 'Dm3tO8L5RItg', 'label': '其他'}]}, 'value': {'value': ['XLyjhSereJYK']}}, {'id': '1yLzipGL8CXc', 'title': '你住在哪~~~', 'kind': 'short_text', 'value': '臺中市'}, {'id': 'umigMKtao9y7', 'title': '來找找適合你的領域~', 'kind': 'multiple_choice', 'properties': {'allowMultiple': True, 'choices': [{'id': 'HDDzjo46UCML', 'label': '公關組'}, {'id': 'BryEnikchCiP', 'label': '活動企劃組'}, {'id': 'QTBS6ThhjimL', 'label': '美術組'}, {'id': 'CL3C1UBMmCpy', 'label': '資訊組'}, {'id': 'LMAYGT8Xo4zc', 'label': '影音組'}]}, 'value': {'value': ['HDDzjo46UCML', 'BryEnikchCiP', 'CL3C1UBMmCpy']}}, {'id': '3DqecY2ogvR5', 'title': '排一排，告訴我們你的優先選擇吧！✨', 'kind': 'long_text', 'value': '第一順位：公關\n第二順位：活動\n第三順位：資訊'}, {'id': 'fV33uN18Aq5b', 'title': '為什麼選擇這些組別？🎯', 'kind': 'long_text', 'value': '有興趣'}, {'id': 'bdXbCUl1iiqq', 'title': '有什麼相關經驗或技能嗎？💡', 'kind': 'long_text', 'value': '嗨嗨嗨嗨嗨嗨嗨嗨 ！\n我是來自臺中二… 不對 我現在沒有學校了😭\n\n\n去年晚些時候我申請了非學校型態實驗教育，在今年一月的時候我成功逃離了二中，在當下雖然很猶豫很迷茫，但現在我很慶幸我逃離了傳統體制！！\n\n\n至於我為什麼逃離二中\n這就是另一個故事了🃏\n\n\n大家可以叫我洪承佑 對… 不是何承祐，不是洪承祐，是洪承佑！！！ 我的名字被打錯ñ次了🙃🙃🙃如果記不起來可以叫我小波！而小波這個名字的由來又是另一個故事了...🃏\n\n\n我是一隻會到處飛來飛去的脆皮小蝴蝶infp，可能上一秒還在快樂的飛來飛去，下一秒掉在地上emo嚎啕大哭（內心劇場。跟我聊天會發現我的思維很跳躍！！！上一秒還在今天晚吃什麼，下一秒你可能會聽到一篇和問題毫不相干的散文？🧊\n\n\n我什麼都可以聊！！沒有雷點！☘️☘️☘️ 但是，但是，不要講話講一半忽然就不講了T_T 我會整天在想什麼啊 什麼啊啊啊啊啊 你為什麼不講了！！！！ 你要講什麼！！！！ 我會因為這內耗一整天的😭😭 有個壞小狗(enfp)曾經這麼做過🫠 還在我好不容易在我忘掉這件事後故意提起來！！！！！ 我真想&_$$#-+（@\n\n\n我其實是彰化人，為了到台中唸書才來到臺中租房子的（（沒錯！我現在自己一個人住😍）），在自學後我開始積極參加比賽和很多 很多 很多課外活動！例如獨旅、志工、黑客松、講座… 也認識了很多很多很好的朋友！我非常喜歡現在的生活，我也很感謝我的家人願意支持我走不一樣的路！！\n\n\n我的專長是資訊/創客領域！未來希望走特殊選材上頂大，目前身上背著五個比賽同時進行中…加小論文六個😭😭😭如果你/妳也喜歡這方面的東西歡迎！！！！！一起參加比賽！！！！\n\n\n最後最後！如果有急事要找我直接傳賴私訊給我！如果非常非常緊急拜託奪命連環call給我😭我的手機24小時開著勿擾，只有賴的私訊是關掉的，其他日常可以用欸居聊就好！\n\n\n很感謝大家願意願意來一起籌備Counterspell_Taiwan，這是台灣第一例由青少年為青少年舉辦黑客松的先例，但其實，早在其他先進國家（如美國、歐洲）由高中生主導的黑客松已經發展多年，而我們希望將這種「青少年為青少年」的理念帶入台灣，並且鼓勵更多人站出來舉辦屬於自己的黑客松。我信這也會對未來的科技人才培育和社會創新產生深遠的影響，為台灣的教育和創新體系帶來新的思維和模式，而不是僅僅限於教科書上的知識，總之！還是很謝謝各位願意給Counterspell Taiwan一個機會！\n（（在此深深一鞠躬？））'}, {'id': 'Gkg271OtI9Bw', 'title': '是否同意我們蒐集並使用您的資料？', 'kind': 'signature', 'value': 'http://form.xiaobo.tw/static/upload/gTB2ITF2CNjG-blob'}], 'hiddenFields': []}
-
-    # 將資料寫入 JSON 檔案
-    json_filename = 'data.json'
-    with open(json_filename, 'w', encoding='utf-8') as json_file:
-        json.dump(data, json_file, ensure_ascii=False, indent=4)
-
-    # 建立 cURL 指令
-    curl_cmd = f'curl -X POST "http://106.107.249.154/webhook" -H "Content-Type: application/json" -d @{json_filename}'
-
-    # 輸出 cURL 指令，或者直接執行系統命令
-    print(f"Generated cURL Command: {curl_cmd}")
-
-    # 執行 cURL 命令
-    os.system(curl_cmd)
-
-    time.sleep(1)
-
