+++ run.py
@@ -7,13 +7,13 @@ import os
 app = create_app()
 
 
-def run_discord_bot():
-    bot.run(os.getenv('DISCORD_TOKEN'))
+# def run_discord_bot():
+#     bot.run(os.getenv('DISCORD_TOKEN'))
 
 
 if __name__ == "__main__":
-    t1 = threading.Thread(target=run_discord_bot)
-    t1.start()
+    # t1 = threading.Thread(target=run_discord_bot)
+    # t1.start()
 
     host = os.getenv('HOST', '0.0.0.0')
     port = int(os.getenv('PORT', 80))
